chore:#33-component  정리
