+++ src/components/Product/ProductCard.jsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import { GoHeart } from 'react-icons/go';
 import { GoHeartFill } from 'react-icons/go';
 import { useRecoilState } from 'recoil';
-import { postsState } from '../recoil/RecoilWastes';
-import { updatePost } from '../api/WastesApi';
+import { postsState } from '../../recoil/RecoilWastes';
+import { updatePost } from '../../api/WastesApi';
 import { Link } from 'react-router-dom';
 const ProductCard = ({ post, onDelete }) => {
   const handleDeleteClick = () => {
